requirements.txt
picamera2==0.3.34
torch== 2.10.0
opencv-python==4.13.0.92
gpiozero==2.0.1
ultralytics==8.4.23
python-dotenv==1.2.2
azure-iot-device==2.14.0